crawler/requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
colored==1.3.93
colorlog==4.0.2
configparser==3.7.4
crayons==0.2.0
idna==2.8
lxml==4.3.4
paho-mqtt==1.4.0
requests==2.18.4
selenium==3.8.0
soupsieve==1.9.1
urllib3==1.22
webdriver-manager==1.7
